removed AngleConverterBase
AngleConverterBase was removed, the needed methods were moved to
OmtfAngleConverter. A few methods that were not used were dropped.

OmtfPhase2AngleConverter
method getGlobalEta renamed to getGlobalEtaPhase2, to avoid hiding
OmtfAngleConverter methods getGlobalEta.
