+++ L1Trigger/L1TMuonOverlapPhase2/plugins/L1TMuonOverlapPhase2TrackProducer.h
@@ -29,7 +29,7 @@ private:
 
   edm::ESGetToken<L1TMuonOverlapParams, L1TMuonOverlapParamsRcd> omtfParamsEsToken;
 
-  //needed for AngleConverterBase
+  //needed for OmtfAngleConverter
   MuonGeometryTokens muonGeometryTokens;
 
   ///needed by tools/CandidateSimMuonMatcher.h
